Upload solution for problem 2
Create class Car: id, Make, Model, Year of manufacture, Color, Price, 
Registration number. Create an array of objects. Save to the different files: 
a) a list of cars of a given brand; 
b) a list of cars of a given model that have been in use for more than n years; 
c) a list of cars of a given year of manufacture, the price of which is higher than the specified one